:ok_hand: (We updated the windows for the service manager user.) We changed the responsive design of the Navbar and Footer
We changed the responsive design of the Navbar and Footer because they weren't working properly in landscape mode. Additionally,
some buttons weren't functioning correctly. We also changed the numbered fonts to more descriptive ones.

BREAKING CHANGE: We changed the responsive design of the Navbar and Footer because they weren't working properly in landscape
mode. Additionally, some buttons weren't functioning correctly. We also changed the numbered fonts to more descriptive ones. The
word "trabajo" was changed to "prevencion" and "extensión" to "cendios."

Closes #1