:iphone: (View on small amd big) Regularize styles for mobile and PC view
Regularize styles for mobile and PC view. Edges and shadows are also added to the head
retainer user, which will later be those introduced in the other roles

BREAKING CHANGE: Regularize styles for mobile and PC view. Edges and shadows are also added
to the head retainer user, which will later be those introduced in the other roles. There is
a lack of views and creating the functionalities of the Firefighters