:white_check_mark: (We perform tests to deliver) We deleted the database and created a new feature.
We deleted the database and created a new feature so that users can be added from a .txt file via the administrator
panel.

BREAKING CHANGE: The tests are positive and it is sent to be evaluated by the application evaluator; it is delivered
with all users empty except for the administrator's bibi@gmail.com, which is a made-up user.

Closes #1